Move dbresources up from plugins to the top-level directory.
