+++ mock/datasync/datasync_mock.go
@@ -20,8 +20,8 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/ligato/cn-infra/datasync"
 
+	"github.com/contiv/vpp/dbresources"
 	controller "github.com/contiv/vpp/plugins/controller/api"
-	"github.com/contiv/vpp/plugins/dbresources"
 )
 
 // MockDataSync can be used to generate datasync events from provided data.
